+++ CHANGELOG.md
@@ -16,6 +16,20 @@ This project adheres to [Semantic Versioning](http://semver.org/) and [Keep a Ch
 ### Breaks
 
 
+## 1.0.20 (2024-04-28)
+---
+
+### New
+Add AWS_SSL_CERTIFICATE_ARN env. var. for those cases where the SSL certificate cannot be found automatically in AWS certificate manager [FA-98].
+
+### Changes
+Change "npm_publish.sh" to show the current date/time at the end [FA-83].
+Change "error-and-reenter.jsx" to show debug info in "getErrorMessage" [FA-246].
+
+### Fixes
+Fix the "ACM certificate ARN not found" error in "aws_deploy_to_s3.sh" and "aws_get_ssl_cert_arn.sh" by using AWS_SSL_CERTIFICATE_ARN env. va. [FA-98].
+
+
 ## 1.0.19 (2024-04-20)
 ---